pyproject.toml
[tool.poetry]
name = "artifact"
version = "0.1.0"
description = ""
authors = ["SkySock <inikitenko@list.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.0"
djangorestframework = "^3.13"
Markdown = "^3.3.6"
psycopg2 = "^2.9.3"
PyJWT = "^2.3.0"
django-cors-headers = "^3.11.0"
Pillow = "^9.1.1"
httplib2 = "^0.20.4"
drf-spectacular = "^0.22.1"
django-money = "^2.1.1"
six = "^1.16.0"
celery = "^5.2.7"
PyAMQP = "^0.1.0"
SQLAlchemy = "^1.4.40"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
